infrastructure/requirements.txt
aws-cdk-lib>2.0,<3.0
chalice==1.29.0